added animation and progress bar instead of entry
